Combine the FeedUserResponse and ListResponse into a single abstract class.
