--- EmbeddedSocialClient/sdk/src/main/java/com/microsoft/embeddedsocial/server/model/FeedUserResponse.java
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See LICENSE in the project root for license information.
- */
-
-package com.microsoft.embeddedsocial.server.model;
-
-/**
- *
- */
-public class FeedUserResponse {
-
-    private String continuationKey;
-
-    public String getContinuationKey() {
-        return continuationKey;
-    }
-
-    public void setContinuationKey(String continuationKey) {
-        this.continuationKey = continuationKey;
-    }
-
-}
