+++ EmbeddedSocialClient/sdk/src/main/java/com/microsoft/embeddedsocial/server/model/ListResponse.java
@@ -8,12 +8,21 @@ package com.microsoft.embeddedsocial.server.model;
 import java.util.List;
 
 /**
- * interface for server responses returning a list of items.
+ * Base class for server responses returning a list of items.
  *
  * @param <T> item type
  */
-public interface ListResponse<T> {
-	List<T> getData();
+public abstract class ListResponse<T> {
 
-	String getContinuationKey();
+	private String continuationKey;
+
+	public String getContinuationKey() {
+		return continuationKey;
+	}
+
+	public void setContinuationKey(String continuationKey) {
+		this.continuationKey = continuationKey;
+	}
+
+	public abstract List<T> getData();
 }
