+++ EmbeddedSocialClient/sdk/src/main/java/com/microsoft/embeddedsocial/server/model/content/replies/GetReplyFeedResponse.java
@@ -6,14 +6,13 @@
 package com.microsoft.embeddedsocial.server.model.content.replies;
 
 import com.microsoft.embeddedsocial.autorest.models.FeedResponseReplyView;
-import com.microsoft.embeddedsocial.server.model.FeedUserResponse;
 import com.microsoft.embeddedsocial.server.model.ListResponse;
 import com.microsoft.embeddedsocial.server.model.view.ReplyView;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class GetReplyFeedResponse extends FeedUserResponse implements ListResponse<ReplyView> {
+public class GetReplyFeedResponse extends ListResponse<ReplyView> {
 
 	private List<ReplyView> replies;
